Create TrackWithBounty
with tracking you don't have to worry much about the wall movement and such. It will follow a robot and stay a distance away from it. It might bump into other robots and the original Tracker had the bot switch to tracking whichever bot it bumped into. I didn't put that behavior in but if you think that's better the code is in the Tracker bot in the sample package in the onHitRobot event. I figured we would track whichever robot had the bigger bounty. So far whenever I've run it, walls are not a problem, but other robots that it runs into are. Walls might be an issue once ramming is implemented (if we miss and hit the wall), but if the others stay far enough from the wall that shouldn't be an issue. We can check position before ramming to see that they are far enough from the border.